Optimized the Cortex-M subregion algorithm
Combined an old idea with the current algorithm. The algorithm checks for every possible subregion size, and tries to make them work while minimally changing start. If not, size is rounded to a power of two and start is simply rounded up so start % size = 0. I'm pretty sure this algorithm find the optimal subregion alignment, tested for a ton of values in the branch memory_layout_danilo on the dverhaert/tock repo.